kp25tk/: Add SWLED to indicate when switch is on/up
We change the pull-up for the power LED from R12 to R4, so it matches
the diode number D4, and change the value to D4 to "PWR." (It's not
clear to me that this is the correct attribute to change to name the
LED "PWR," though.)

We add LED D5 and pull-up R5 driving it, pulled down by /SWITCH
(negative logic output of slide switch buffers) to light it when the
switch is on.

README is updated with the design notes for this, and also a new
component summary (only partially complete) which serves as a rough
guide to the reference numbering we're using on components.