romext-2364 sch: Complete schematic of basic version
The "2364" is actually a socket J1, where the addr/data/select
lines are outputs because this is wired to the socket on the
PC-8001 motherboard.

This is the simplest design that will work with the PC-8001 and
an ST Microelectronics M2764A; it does not yet provide facilities
to work with systems or EPROMs that vary how they use chip select
or have different requirements for levels on the programming pins
during read operations.