BROKEN Test: Gerber generation for romext-2364
XXX generates new files every time due to creation date being updated.
Fix this by maybe generating to temp dir, diffing with committed files,
and not updating if creation date is the only thing that changed?

This is actually generic to all projects, but while we're testing out
how well this works (including getting boards fabricated and
inspecting them), we'll do it just for the romext-2364 project.