render: Generate PDF schematics and Gerber PCB layouts
This is currently run only manually because even if there have been
no changes to the source files, it always generates changed output
due to the creation date in the files. (This needs to be fixed before
we start doing this automatically in `Test`.)

Test no longer does the automatic schematic rendering; we're waiting
for `render` to be fixed.