romext-2364 pcb: Change U1/2764 to use standard library ZIF footprint
Bumps PCB revision to 1-2dev.

The footprint was hard to find[1]; one has to turn off the footprint
filter to have it come up.

This updates the schematic as well, but just the footprint assignment
in it, so we don't need to bump the revision. (It does update the PDF,
however, because that contains the footprint info in the internal
Javascript for popups or whatever.)

[1]: https://forum.kicad.info/t/please-add-zif-to-the-descriptions-of-zero-insertion-force-sockets/62549/6