rom-read-adapter sch: ZIF symbol update from lib and footprint; pin updates
The pin updates are for JR3 (the pin header connected directly to ZIF
pins 25-48) where we had pins numbered from 25-48 on JR3 also numbered
to match. Not clear if this was a problem, because we also had the
connections between the ZIF and this header in a real mess, but it's
now exactly pin-to-pin as it should always have been. Before we
weren't seeing the correct connections on the PCB editor ratsnest; now
we are.