render: Fix main loop to work with multiple arguments
The main loop was mostly still using $1 even though $arg was the
current argument it was working on, so this caused it to render just
the first argument over and over again.

Also now include the file extension in the "───── Rendering ..."
output so that we can more easily see which rendering function is
being used.