render: Use strip-nondeterminism on $pcbname-jlcpcb.zip
We always re-create $pcbname-jlcpcb.zip; this will produce a version
with different timestamps every time it's run. strip-nondeterminism
undoes the timestamps (both for archive creation and the files within
it) so that when it's re-created from the same source files, the same
archive is produced.

This has the slightly unfortunate side-effect of setting all the file
modification dates in the archive to 1980-01-01 00:00 UTC, but a better
solution to that (keeping the last modification times unchanged from
the previous archive) appears to be quite difficult.

This also now lets the error message show when we check if `kicad-cli`
exists, just in case the problem is something other than "command not
found."