kp25tk pcb: Remove all stuff below the bottom cut edge
Turns out that, though in the Gerber viewer on JLCPCB we seemed to be
getting what we wanted, it bases the board size not on the edge cuts
but on everything that's on the board, so though the Gerber output
showed an ~100×100 board, we were being charged for a 100×130 or
whatever it was.

This gets the board price down, but unfortunately it still seems too
big for $1 OCS NEP shipping, causing us to have to use the $10 shipping
instead. So we need to look at whether we can cut this down further.