WIP rom-read-adapter pcb: Everything routed except bypass cap
We need to figure out the best physical position for the bypass cap.
Ideally it should be as close as possible to the device pins
themselves, but we don't know which device pins are power; those are
determined half way across the board in the work area. So maybe place
it across the pin 12/24 GND and pin 24/24 VCC? Probably wants to be an
axial in that case, perhaps positioned top-to-bottom in the RHS
between the default jumpers and the VCC/CS/¬CS/GND jumpers.