WIP rom-read-adapter: Replace disc cap with axial; all traces routed
We've gone with the solution mentioned in the previous commit message,
where we put an axial cap kinda-near the GND and VCC pins in the work
area. It could be closer, by moving some traces, or even by putting it
in between the socket pin connectors, though that would cover up lots
of pin numbers. (Maybe try it on the bottom? Still would moving some
traces to make room for the pads.)