kicad-lib: Header_Pin and ZIF48 symbols; DIP_Socket-48 footprint
This is the symbol and footprint for a 48-pin ZIF socket,
"DIP_Socket-48_W11.9_W12.7_W15.24_W17.78_W18.5",
based on the 3M 42-pin ZIF socket in the standard library,
"DIP_Socket-42_W11.9_W12.7_W15.24_W17.78_W18.5_3M_242-1281-00-0602J".

Probably the W11.9 etc. should have 'mm' after them. Though the
existing footprints for ZIF sockets do not have that, this sems to be
a bug that has been fixed in later versions of the standard library.
(Without the 'mm' suffix they don't match the footprint filters for
certain symbols that could use this socket.)

The Header_Pin symbols are an experiment that we decided not to use
in an upcoming project; we should examine them and see if we really
need to keep them.