kp25tk pcb: Connect all endpoints of board edge
The board edge had line and curve endpoints that didn't quite match up
with each other, so it wasn't a proper closed shape. Apparently JLCPCB
had no problem with this (except that removed the curve from the upper
right hand corner and made it 90° instead), but the tracespace.io
viewer made most of the board vanish when "Use outline layer for board
shape" is checked in the render options, which is what alerted us
to the problem.

We've changed each start point to match the previous segment's end
point and confirmed that all lines are going clockwise to make the
enclosed shape. There ought to be a way to have the DRC check do this,
really.

Bump version from 0.0.8 to 0.0.9, but we should switch this to
use our standard version numbering before bringing this on to main.